Add colorblind mode to Mineseeker
Also tried to add image functionality to Mineseeker for the analyzer. It works, but the bullet positions are bugging me, and the log is being read in the wrong order due to parsing the functions after the text is applied. (In other words, the functions are done last)

Improvements to the new table key and overall HTML are welcomed.