Index parseData before access
Create a separate array that indexes the fields that getModuleData() is referencing, so it need not iterate over the entirety of parseData every single time. This was the single biggest bottleneck I could find by profiling the LFA while it analyzed a large log file, and making this changed dropped the load time for a ~6mb log file from almost two minutes, to about five seconds.