Russian Translations (#1055)
* Fixed russian Shape Shift manual and translated some manuals

* - Merged my Shape Shift with the actual manual (excuse me for making a second manual)
- Fixed tables and included cyrillic for russian Button Sequence

* -Not Keypad translation

* -Algebra table fix
-Juxta Squares fixes
-Deleted now redundant Shape Shift manual(the one that i created)

* Translated Visual Impairment

* Translated Big Circle
(i'm not sure if translated ruleseed will work though, cuz i can't check it)

* Translated Waste Management

* Web Design table fix
Not really a pro in html, but it seems that including cyrillic css solves 95% of problems with tables

* Text Field text fix
+cyrillic css, yay

* Crazy Talk table fix
well, not really, but at least it doesn't leave the page

* Bitwise operations table enhancement
+cyrillic css

* Color math table fix
cyrillic css is a cure to all problems

* Added translated SVGs to Complicated Buttons

* Translated Digital Root

* Translated The Code

* Little Not Keypad translation fix
Will Will Smith translate itself into russian? Or maybe i should pay more attention to translations?

* Fixed linter issues

* Added cyrillic css to Silly slots; Translated Burglar Alarm

* Added link to russian juxta squares in colored squares manual

* Translated Uncolored Switches
guess this manual will do for now, not really sure this will work. Oh well

* Added identification appendix to russian colored switches manual
I'll add apropriate links a bit later(when there will be such links)

* Translated recolored switches and added a link to colored switches

* Translated egg
Because someone made a joke about it

* Joke update - an egg svg

* Translated multicolored switches and finished identification section in colored switches
also i just realised that i dont have to refer to the site in links. I'll leave the colored squares be as it is for now

* Translated mazematics + grammar fixes

* Translated Forget It Not

* egg svg fix

* more minor fixes

* thanks english, i forgot that "голубой" exists

Co-authored-by: shortc1rcuit <declan@themcmoos.net>