Add developer mode to the in-manual options menu
This mode is intended for maintainers and manual creators to use when checking manuals before pushing them to the repo.
Currently, all this does is colour the background red, which will only be visible if the manual exceeds its page boundaries.
For this reason, the entry in the options menu, for now, is "Highlight overflow".