Russian addendum to the quotation event
First of all - for the love of god, please consult with either the russian community or russian maintainers before submitting such large changes to the translations.

Secondly - on the subject of arrow quotes.
While arrow quotes are recognized by our goverment and typographers as "the right quotes" I strongly disagree with this replacement.
Arrow quotes appear in our media and legal/offical documents because they are written in an editor that automatically replaces double "" quotes with the arrow quotes.

The "" quotes are more familiar to the average russian when using digital media. Hint: DIGITAL. None of russian keyboard layouts have arrow quotes.
The only way to write them is by using alt codes, which only works with a numpad (which some of us simply do not have). In the context of manuals it's also possible to use HTML codes, but you can't expect an average person to keep in mind that they should be using goddamn codes instead of usual Shift+2. It's very inconvenient.

This replacement is not worth the "style points" it would get, because it's just not worth the hustle.